Array DB helpers: * Remove any item from the arraydb * Get items from certain index in array db
